move workspace-action files into a shared folder
