+++ src/packages/core/workspace/components/workspace-action/index.ts
@@ -1,3 +1,2 @@
-export * from './workspace-action-base.js';
-export * from './workspace-action.element.js';
+export * from './shared/index.js';
 export * from './common/index.js';
