+++ src/packages/core/workspace/components/workspace-action/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './workspace-action/index.js';
+export * from './workspace-action-menu/index.js';
